Create CNN 1D classification of press-in curves
This code was a part of my bachelor paper. 
I analyzed in this paper a 1-dimensional press-in curves that I obtained from a press-in experiment. 
The goal of this work was to classify several press-in conditons with CNN 1D.
So you can check out how to classify a 1D data with CNN within 9 steps.